Ensure that tests for masking REST Client headers actually run
The previous version of the tests didn't run at all so it
presented a false sense of security
